+++ src/layers/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod pip;
 pub(crate) mod pip_cache;
 pub(crate) mod pip_dependencies;
 pub(crate) mod python;